Make Rails app detection based on Rails::Application superclass
This change makes the Rails app detection based on the superclass of the
main class being `Rails::Application`. This is a more reliable way to
detect Rails apps than looking for the presence of the `rails` gem in
the Gemfile. Application can require some of the Rails components
without requiring the `rails` gem itself.

co-authored-by: Earlopain <14981592+Earlopain@users.noreply.github.com>