Fix spec code lens inside classes
Minitest specs don't include the class name in the test selector, so we
need to strip out the class name (including the module nesting) when we
generate the selector.

This commit changes `@group_stack` to be a tuple of `[type, value]` so
we can distinguish between class/module names and test group/example
names. For specs, it removes the class/module names from the stack.